% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_compact}
\alias{h3_compact}
\title{Compact}
\usage{
h3_compact(h3Set)
}
\arguments{
\item{h3Set}{Array.<H3Index> H3 indexes to compact}
}
\value{
\code{Array.<H3Index>} - Compacted H3 indexes
}
\description{
Compact a set of hexagons of the same resolution into a set of hexagons acrossmultiple levels that represents the same area.
}
\details{
Throws \code{Error} If the input is invalid (e.g. duplicate hexagons)
}
