% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_degs_to_rads}
\alias{h3_degs_to_rads}
\title{Degs to rads}
\usage{
h3_degs_to_rads(deg)
}
\arguments{
\item{deg}{Number Value in degrees}
}
\value{
\code{Number} - Value in radians
}
\description{
Convert degrees to radians
}
