% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_distance}
\alias{h3_distance}
\title{H3 distance}
\usage{
h3_distance(origin, destination)
}
\arguments{
\item{origin}{H3Index Origin hexagon index}

\item{destination}{H3Index Destination hexagon index}
}
\value{
\code{Number} - Distance between hexagons, or a negative                              number if the distance could not be computed
}
\description{
Get the grid distance between two hex indexes. This function may failto find the distance between two indexes if they are very far apart oron opposite sides of a pentagon.
}
