% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_edge_length}
\alias{h3_edge_length}
\title{Edge length}
\usage{
h3_edge_length(res, unit)
}
\arguments{
\item{res}{Number Hexagon resolution}

\item{unit}{String Area unit (either UNITS.m or UNITS.km)}
}
\value{
\code{Number} - Average edge length
}
\description{
Average hexagon edge length at a given resolution
}
\details{
Throws \code{Error} If the unit is invalid
}
