% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_get_base_cell}
\alias{h3_get_base_cell}
\title{H3 get base cell}
\usage{
h3_get_base_cell(h3Index)
}
\arguments{
\item{h3Index}{\code{H3Index} H3 index to get the base cell for}
}
\value{
\code{Number} - Index of the base cell (0-121)
}
\description{
Get the number of the base cell for a given H3 index
}
