% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_get_destination_h3_index_from_unidirectional_edge}
\alias{h3_get_destination_h3_index_from_unidirectional_edge}
\title{Get destination h3 index from unidirectional edge}
\usage{
h3_get_destination_h3_index_from_unidirectional_edge(edgeIndex)
}
\arguments{
\item{edgeIndex}{H3Index H3 index of the edge}
}
\value{
\code{H3Index} - H3 index of the edge destination
}
\description{
Get the destination hexagon from an H3 index representing a unidirectional edge
}
