% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_get_h3_indexes_from_unidirectional_edge}
\alias{h3_get_h3_indexes_from_unidirectional_edge}
\title{Get h3 indexes from unidirectional edge}
\usage{
h3_get_h3_indexes_from_unidirectional_edge(edgeIndex)
}
\arguments{
\item{edgeIndex}{\code{H3Index} H3 index of the edge}
}
\value{
\code{Array.<H3Index>} - \code{[origin, destination]} pair as H3 indexes
}
\description{
Get the \code{[origin, destination]} pair represented by a unidirectional edge
}
