% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_get_h3_unidirectional_edge}
\alias{h3_get_h3_unidirectional_edge}
\title{Get h3 unidirectional edge}
\usage{
h3_get_h3_unidirectional_edge(origin, destination)
}
\arguments{
\item{origin}{\code{H3Index} Origin hexagon index}

\item{destination}{\code{H3Index} Destination hexagon index}
}
\value{
\code{H3Index} - H3 index of the edge, or null if no edge is shared
}
\description{
Get an H3 index representing a unidirectional edge for a given origin and destination
}
