% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_get_h3_unidirectional_edge_boundary}
\alias{h3_get_h3_unidirectional_edge_boundary}
\title{Get h3 unidirectional edge boundary}
\usage{
h3_get_h3_unidirectional_edge_boundary(edgeIndex,
  formatAsGeoJson = FALSE)
}
\arguments{
\item{edgeIndex}{\code{H3Index} H3 index of the edge}

\item{formatAsGeoJson}{\code{Boolean} Whether to provide GeoJSON output: \code{[lng, lat]}}
}
\value{
\code{Array.<Array>} - Array of geo coordinate pairs
}
\description{
Get the vertices of a given edge as an array of \code{[lat, lng]} points. Note that for edges thatcross the edge of an icosahedron face, this may return 3 coordinates.
}
