% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_get_h3_unidirectional_edges_from_hexagon}
\alias{h3_get_h3_unidirectional_edges_from_hexagon}
\title{Get h3 unidirectional edges from hexagon}
\usage{
h3_get_h3_unidirectional_edges_from_hexagon(h3Index)
}
\arguments{
\item{h3Index}{H3Index H3 index of the origin hexagon}
}
\value{
\code{Array.<H3Index>} - List of unidirectional edges
}
\description{
Get all of the unidirectional edges with the given H3 index as the origin (i.e. an edge toevery neighbor)
}
