% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_get_resolution}
\alias{h3_get_resolution}
\title{H3 get resolution}
\usage{
h3_get_resolution(h3Index)
}
\arguments{
\item{h3Index}{\code{H3Index} H3 index to get resolution}
}
\value{
\code{Number} - The number (0-15) resolution, or -1 if invalid
}
\description{
Returns the resolution of an H3 index
}
