% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_hex_area}
\alias{h3_hex_area}
\title{Hex area}
\usage{
h3_hex_area(res, unit)
}
\arguments{
\item{res}{\code{Number} Hexagon resolution}

\item{unit}{\code{String} Area unit (either UNITS.m2 or UNITS.km2)}
}
\value{
\code{Number} - Average area
}
\description{
Average hexagon area at a given resolution
}
\details{
Throws \code{Error} If the unit is invalid
}
