% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_hex_ring}
\alias{h3_hex_ring}
\title{Hex ring}
\usage{
h3_hex_ring(h3Index, ringSize)
}
\arguments{
\item{h3Index}{H3Index H3 index of center hexagon}

\item{ringSize}{Number Radius of ring}
}
\value{
\code{Array.<H3Index>} - H3 indexes for all hexagons in ring
}
\description{
Get all hexagons in a hollow hexagonal ring centered at origin with sides of a given length.Unlike kRing, this function will throw an error if there is a pentagon anywhere in the ring.
}
\details{
Throws \code{Error} If the algorithm could not calculate the ring
}
