% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_indexes_are_neighbors}
\alias{h3_indexes_are_neighbors}
\title{H3 indexes are neighbors}
\usage{
h3_indexes_are_neighbors(origin, destination)
}
\arguments{
\item{origin}{H3Index Origin hexagon index}

\item{destination}{H3Index Destination hexagon index}
}
\value{
\code{Boolean} - Whether the hexagons share an edge
}
\description{
Whether two H3 indexes are neighbors (share an edge)
}
