% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_is_pentagon}
\alias{h3_is_pentagon}
\title{H3 is pentagon}
\usage{
h3_is_pentagon(h3Index)
}
\arguments{
\item{h3Index}{\code{H3Index} H3 index to check}
}
\value{
\code{Boolean} - isPentagon
}
\description{
Whether the given H3 index is a pentagon
}
