% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_is_res_class_iii}
\alias{h3_is_res_class_iii}
\title{H3 is res class iii}
\usage{
h3_is_res_class_iii(h3Index)
}
\arguments{
\item{h3Index}{H3Index H3 index to check}
}
\value{
\code{Boolean} - isResClassIII
}
\description{
Whether the given H3 index is in a Class III resolution (rotated versusthe icosahedron and subject to shape distortion adding extra points onicosahedron edges, making them not true hexagons).
}
