% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_is_valid}
\alias{h3_is_valid}
\title{H3 is valid}
\usage{
h3_is_valid(h3Index)
}
\arguments{
\item{h3Index}{\code{H3Index} H3 index to check}
}
\value{
\code{Boolean} - Whether the index is valid
}
\description{
Whether a given string represents a valid H3 index
}
