% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_k_ring}
\alias{h3_k_ring}
\title{K ring}
\usage{
h3_k_ring(h3Index, ringSize)
}
\arguments{
\item{h3Index}{H3Index H3 index of center hexagon}

\item{ringSize}{Number Radius of k-ring}
}
\value{
\code{Array.<H3Index>} - H3 indexes for all hexagons in ring
}
\description{
Get all hexagons in a k-ring around a given center. The order of the hexagons is undefined.
}
