% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_num_hexagons}
\alias{h3_num_hexagons}
\title{Num hexagons}
\usage{
h3_num_hexagons(res)
}
\arguments{
\item{res}{\code{Number} Hexagon resolution}
}
\value{
\code{Number} - Count
}
\description{
The total count of hexagons in the world at a given resolution. Note that aboveresolution 8 the exact count cannot be represented in a JavaScript 32-bit number,so consumers should use caution when applying further operations to the output.
}
