% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_polyfill}
\alias{h3_polyfill}
\title{Polyfill}
\usage{
h3_polyfill(coordinates, res, isGeoJson = FALSE)
}
\arguments{
\item{coordinates}{\code{Array.<Array>} Array of loops, or a single loop}

\item{res}{\code{Number} Resolution of hexagons to return}

\item{isGeoJson}{\code{Boolean} Whether to expect GeoJson-style \code{[lng, lat]}                                 pairs instead of \code{[lat, lng]}}
}
\value{
\code{Array.<H3Index>} - H3 indexes for all hexagons in polygon
}
\description{
Get all hexagons with centers contained in a given polygon. The polygonis specified with GeoJson semantics as an array of loops. Each loop isan array of \code{[lat, lng]} pairs (or \code{[lng, lat]} if isGeoJson is specified).The first loop is the perimeter of the polygon, and subsequent loops areexpected to be holes.
}
