% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_rads_to_degs}
\alias{h3_rads_to_degs}
\title{Rads to degs}
\usage{
h3_rads_to_degs(rad)
}
\arguments{
\item{rad}{Number Value in radians}
}
\value{
\code{Number} - Value in degrees
}
\description{
Convert radians to degrees
}
