% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_set_to_multi_polygon}
\alias{h3_set_to_multi_polygon}
\title{H3 set to multi polygon}
\usage{
h3_set_to_multi_polygon(h3Indexes, formatAsGeoJson = FALSE)
}
\arguments{
\item{h3Indexes}{\code{Array.<H3Index>} H3 indexes to get outlines for}

\item{formatAsGeoJson}{\code{Boolean} Whether to provide GeoJSON output:                                   \code{[lng, lat]}, closed loops}
}
\value{
\code{Array.<Array>} - MultiPolygon-style output.
}
\description{
Get the outlines of a set of H3 hexagons, returned in GeoJSON MultiPolygonformat (an array of polygons, each with an array of loops, each an array ofcoordinates). Coordinates are returned as \code{[lat, lng]} pairs unless GeoJSONis requested.
}
