% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_to_children}
\alias{h3_to_children}
\title{H3 to children}
\usage{
h3_to_children(h3Index, res)
}
\arguments{
\item{h3Index}{\code{H3Index} H3 index to get children for}

\item{res}{\code{Number} Resolution of hexagons to return}
}
\value{
\code{Array.<H3Index>} - H3 indexes of children, or empty array for invalid input
}
\description{
Get the children/descendents of the given hexagon at a particular resolution
}
