% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_to_geo}
\alias{h3_to_geo}
\title{H3 to geo}
\usage{
h3_to_geo(h3Index)
}
\arguments{
\item{h3Index}{H3Index H3 index}
}
\value{
\code{Array.<Number>} - Point as a \code{[lat, lng]} pair
}
\description{
Get the lat,lon center of a given hexagon
}
