% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_to_geo_boundary}
\alias{h3_to_geo_boundary}
\title{H3 to geo boundary}
\usage{
h3_to_geo_boundary(h3Index, formatAsGeoJson = FALSE)
}
\arguments{
\item{h3Index}{\code{H3Index} H3 index}

\item{formatAsGeoJson}{\code{Boolean} Whether to provide GeoJSON output: \code{[lng, lat]}, closed loops}
}
\value{
\code{Array.<Array>} - Array of \code{[lat, lng]} pairs
}
\description{
Get the vertices of a given hexagon (or pentagon), as an array of \code{[lat, lng]}points. For pentagons and hexagons on the edge of an icosahedron face, thisfunction may return up to 10 vertices.
}
