% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_to_parent}
\alias{h3_to_parent}
\title{H3 to parent}
\usage{
h3_to_parent(h3Index, res)
}
\arguments{
\item{h3Index}{\code{H3Index} H3 index to get parent for}

\item{res}{\code{Number} Resolution of hexagon to return}
}
\value{
\code{H3Index} - H3 index of parent, or null for invalid input
}
\description{
Get the parent of the given hexagon at a particular resolution
}
