% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_uncompact}
\alias{h3_uncompact}
\title{Uncompact}
\usage{
h3_uncompact(compactedSet, res)
}
\arguments{
\item{compactedSet}{\code{Array.<H3Index>} H3 indexes to uncompact}

\item{res}{\code{Number} The resolution to uncompact to}
}
\value{
\code{Array.<H3Index>} - The uncompacted H3 indexes
}
\description{
Uncompact a compacted set of hexagons to hexagons of the same resolution
}
\details{
Throws \code{Error} If the input is invalid (e.g. invalid resolution)
}
