% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_unidirectional_edge_is_valid}
\alias{h3_unidirectional_edge_is_valid}
\title{H3 unidirectional edge is valid}
\usage{
h3_unidirectional_edge_is_valid(edgeIndex)
}
\arguments{
\item{edgeIndex}{H3Index H3 index of the edge}
}
\value{
\code{Boolean} - Whether the index is valid
}
\description{
Whether the input is a valid unidirectional edge
}
