% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h3api.R
\name{h3_geo_to_h3}
\alias{h3_geo_to_h3}
\title{Geo to h3}
\usage{
h3_geo_to_h3(lat, lng, res)
}
\arguments{
\item{lat}{\code{Number} Latitude of point}

\item{lng}{\code{Number} Longtitude of point}

\item{res}{\code{Number} Resolution of hexagons to return}
}
\value{
\code{H3Index} - H3 index
}
\description{
Get the hexagon containing a lat,lon point
}
